+++ bootique-job/src/main/java/io/bootique/job/command/ScheduleCommand.java
@@ -2,24 +2,35 @@ package io.bootique.job.command;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import io.bootique.meta.application.CommandMetadata;
 import io.bootique.cli.Cli;
 import io.bootique.command.CommandOutcome;
 import io.bootique.command.CommandWithMetadata;
 import io.bootique.job.scheduler.Scheduler;
+import io.bootique.meta.application.CommandMetadata;
+import io.bootique.meta.application.OptionMetadata;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.List;
 
 public class ScheduleCommand extends CommandWithMetadata {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(ScheduleCommand.class);
 
+	public static final String JOB_OPTION = "job";
+
 	private Provider<Scheduler> schedulerProvider;
+
+	private static OptionMetadata.Builder createJobOption() {
+		return OptionMetadata.builder(JOB_OPTION).description("Specifies the name of the job to schedule. "
+				+ "Available job names can be viewed using '--list' command.").valueRequired("job_name");
+	}
 
 	private static CommandMetadata createMetadata() {
 		return CommandMetadata.builder(ScheduleCommand.class)
 				.description(
-						"Schedules and executes jobs according to configuration. Waits indefinitely on the foreground.")
+						"Schedules and executes jobs according to configuration and '--job' arguments. Waits indefinitely on the foreground.")
+				.addOption(createJobOption())
 				.build();
 	}
 
@@ -31,13 +42,20 @@ public class ScheduleCommand extends CommandWithMetadata {
 
 	@Override
 	public CommandOutcome run(Cli cli) {
-
 		Scheduler scheduler = schedulerProvider.get();
 
-		// TODO: filter configured triggers by jobs specified with --jobs
+		int jobCount;
 
-		LOGGER.info("Starting scheduler");
-		if (scheduler.start() > 0) {
+		List<String> jobNames = cli.optionStrings(JOB_OPTION);
+		if (jobNames == null || jobNames.isEmpty()) {
+			LOGGER.info("Starting scheduler");
+			jobCount = scheduler.start();
+		} else {
+			LOGGER.info("Starting scheduler for jobs: " + jobNames);
+			jobCount = scheduler.start(jobNames);
+		}
+
+		if (jobCount > 0) {
 			try {
 				Thread.currentThread().join();
 			} catch (InterruptedException e) {