+++ bootique-job/src/main/java/io/bootique/job/scheduler/DefaultScheduler.java
@@ -18,8 +18,12 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
+import java.util.TreeSet;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.BiFunction;
@@ -34,6 +38,7 @@ public class DefaultScheduler implements Scheduler {
     private RunnableJobFactory runnableJobFactory;
     private JobRegistry jobRegistry;
     private Collection<TriggerDescriptor> triggers;
+    private Map<String, Collection<TriggerDescriptor>> triggerMap;
     private Map<String, Collection<ScheduledJobFuture>> scheduledJobsByName;
 
     private AtomicBoolean started;
@@ -45,19 +50,55 @@ public class DefaultScheduler implements Scheduler {
         this.taskScheduler = taskScheduler;
         this.runnableJobFactory = runnableJobFactory;
         this.jobRegistry = jobRegistry;
-        this.triggers = Collections.unmodifiableCollection(triggers);
+        this.triggers = triggers;
+        this.triggerMap = collectTriggers(triggers);
         this.scheduledJobsByName = new HashMap<>();
 
         this.started = new AtomicBoolean(false);
     }
 
+    private Map<String, Collection<TriggerDescriptor>> collectTriggers(Collection<TriggerDescriptor> triggers) {
+        return triggers.stream().collect(
+                Collectors.toMap(
+                        TriggerDescriptor::getJob,
+                        t -> new ArrayList<>(Collections.singleton(t)),
+                        (l1, l2) -> {l1.addAll(l2); return l1;}));
+    }
+
     @Override
     public int start() {
-        if (!started.compareAndSet(false, true)) {
-            throw new IllegalStateException("Already started");
+        return scheduleTriggers(triggers);
+    }
+
+    @Override
+    public int start(List<String> jobNames) {
+        if (jobNames == null || jobNames.isEmpty()) {
+            throw new BootiqueException(1, "No jobs specified");
         }
 
-        if (triggers.isEmpty()) {
+        Set<String> jobNamesSet = new TreeSet<>(jobNames);
+        Iterator<String> iter = jobNamesSet.iterator();
+        while (iter.hasNext()) {
+            String jobName = iter.next();
+            if (!jobRegistry.getAvailableJobs().contains(jobName)) {
+                throw new BootiqueException(1, "Unknown job: " + jobName);
+            } else if (!triggerMap.containsKey(jobName)) {
+                LOGGER.warn("No triggers configured for job: {}. Skipping...", jobName);
+                iter.remove();
+            }
+        }
+
+        Collection<TriggerDescriptor> triggers = jobNamesSet.stream()
+                .flatMap(jobName -> triggerMap.get(jobName).stream())
+                .collect(Collectors.toList());
+
+        return scheduleTriggers(triggers);
+    }
+
+    private int scheduleTriggers(Collection<TriggerDescriptor> triggers) {
+        int triggerCount = triggers.size();
+
+        if (triggerCount == 0) {
             LOGGER.info("No triggers, exiting");
             return 0;
         }
@@ -71,26 +112,37 @@ public class DefaultScheduler implements Scheduler {
             throw new BootiqueException(1, "Trigger(s) without a job object: " + badTriggers);
         }
 
-        triggers.forEach(tc -> {
-            Job job = jobRegistry.getJob(tc.getJob());
-            String jobName = job.getMetadata().getName();
 
-            Function<Schedule, JobFuture> scheduler = (schedule) -> {
-                LOGGER.info(String.format("Will schedule '%s'.. (%s)", jobName, schedule.getDescription()));
-                return schedule(job, Collections.emptyMap(), schedule.getTrigger());
-            };
+        if (triggerCount > 0) {
+            tryStart();
 
-            ScheduledJobFuture scheduledJob = new DefaultScheduledJobFuture(jobName, scheduler);
-            scheduledJob.schedule(createSchedule(tc));
-            Collection<ScheduledJobFuture> futures = scheduledJobsByName.get(jobName);
-            if (futures == null) {
-                futures = new ArrayList<>();
-                scheduledJobsByName.put(jobName, futures);
-            }
-            futures.add(scheduledJob);
-        });
+            triggers.forEach(tc -> {
+                Job job = jobRegistry.getJob(tc.getJob());
+                String jobName = job.getMetadata().getName();
 
-        return triggers.size();
+                Function<Schedule, JobFuture> scheduler = (schedule) -> {
+                    LOGGER.info(String.format("Will schedule '%s'.. (%s)", jobName, schedule.getDescription()));
+                    return schedule(job, Collections.emptyMap(), schedule.getTrigger());
+                };
+
+                ScheduledJobFuture scheduledJob = new DefaultScheduledJobFuture(jobName, scheduler);
+                scheduledJob.schedule(createSchedule(tc));
+                Collection<ScheduledJobFuture> futures = scheduledJobsByName.get(jobName);
+                if (futures == null) {
+                    futures = new ArrayList<>();
+                    scheduledJobsByName.put(jobName, futures);
+                }
+                futures.add(scheduledJob);
+            });
+        }
+
+        return triggerCount;
+    }
+
+    private void tryStart() {
+        if (!started.compareAndSet(false, true)) {
+            throw new IllegalStateException("Already started");
+        }
     }
 
     private Schedule createSchedule(TriggerDescriptor tc) {